+++ packages/react/src/index.js
@@ -10,6 +10,7 @@ export FilterAdder from './FilterAdder'
 export FilterButtonList from './FilterButtonList'
 export SearchFilters from './SearchFilters'
 export SearchLayout from './SearchLayout'
+export ToggleFiltersHeader from './ToggleFiltersHeader'
 
 // component library
 export * from './greyVest'