+++ packages/react/src/purgatory/index.js
@@ -1,4 +1,7 @@
 export CheckButton from './CheckButton'
 export ModalPicker from './ModalPicker'
 export PopoverTagsInput from './PopoverTagsInput'
+export SearchTree from './SearchTree'
 export StepsAccordion, { AccordionStep } from './StepsAccordion'
+export ToggleFiltersButton from './ToggleFiltersButton'
+export TreePauseButton from './TreePauseButton'