+++ packages/react/stories/imdb/stories/greyVestThemed.js
@@ -10,15 +10,10 @@ import {
   componentForType,
   SearchLayout,
   SearchFilters,
-} from '../../../src'
-import {
-  SearchTree,
   ToggleFiltersHeader,
-  Tab,
-  TabContent,
-  TabLabel,
-  Tabs,
-} from '../../../src/greyVest'
+} from '../../../src'
+import { SearchTree } from '../../../src/purgatory'
+import { Tab, TabContent, TabLabel, Tabs } from '../../../src/greyVest'
 import {
   DateRangePicker,
   TypeMap,